attempting to put children inside circles
I can create a child in a circle, but my selection method returns a node
and the child is a circle. I think i need to create a new linklist
inside each node when I add children. The big problem is figuring out
how to constrain the circles/nodes inside the bigger ones.