Added children and scrollwheel to change radii
bugs left with children:

their bound detection is slightly off, so they can sometimes escape the
larger circle in the corners.

Also if you drag a stack of small circles off of a larger circle, the
children of the one that is selected will keep following the mouse even
outside of the bounds. However, they snap back to their parent circle as
soon as any circle in the family is moved.